pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-rag-server"
version = "2.1.5"
description = "Serveur MCP avec RAG complet sur codebase Python/C++ (chunking syntaxique, ChromaDB, mistral-embed)"
requires-python = ">=3.11"

dependencies = [
    # MCP
    "mcp>=1.0.0",

    # LangChain + Mistral
    "langchain-openai>=0.1",
    
    # ChromaDB (base vectorielle avec persistance SQLite)
    "chromadb>=0.5",

    # Chunking C++
    "tree-sitter>=0.23",
    "tree-sitter-cpp>=0.23",
    "pymupdf>=1.24",

    # Utilitaires
    "python-dotenv>=1.0",

    "pydantic<3.0.0,>=2.11.0",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
    "jupyter",
]

# Indique à setuptools que le projet est un répertoire de scripts plats
# (pas de package Python structuré avec __init__.py)
[tool.setuptools]
py-modules = [
    "chunker",
    "embedder",
    "indexer",
    "mcp_rag_server",
    "mcp_client_llm",
    "pdf_chunker",
    "retriever",
    "store",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
asyncio_mode = "auto"